move type translators to common package
