+++ pkg/services/authz/zanzana/zanzana.go
@@ -7,7 +7,6 @@ import (
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
 	"github.com/grafana/grafana/pkg/services/authz/zanzana/common"
-	authzextv1 "github.com/grafana/grafana/pkg/services/authz/zanzana/proto/v1"
 )
 
 const (
@@ -121,85 +120,4 @@ func TranslateFixedRole(role string) string {
 // Translate "read" for the dashboard into "dashboard_read" for folder
 func TranslateToFolderRelation(relation, objectType string) string {
 	return fmt.Sprintf("%s_%s", objectType, relation)
-}
-
-func ToAuthzExtTupleKey(t *openfgav1.TupleKey) *authzextv1.TupleKey {
-	tupleKey := &authzextv1.TupleKey{
-		User:     t.GetUser(),
-		Relation: t.GetRelation(),
-		Object:   t.GetObject(),
-	}
-
-	if t.GetCondition() != nil {
-		tupleKey.Condition = &authzextv1.RelationshipCondition{
-			Name:    t.GetCondition().GetName(),
-			Context: t.GetCondition().GetContext(),
-		}
-	}
-
-	return tupleKey
-}
-
-func ToAuthzExtTupleKeys(tuples []*openfgav1.TupleKey) []*authzextv1.TupleKey {
-	result := make([]*authzextv1.TupleKey, 0, len(tuples))
-	for _, t := range tuples {
-		result = append(result, ToAuthzExtTupleKey(t))
-	}
-	return result
-}
-
-func ToAuthzExtTupleKeyWithoutCondition(t *openfgav1.TupleKeyWithoutCondition) *authzextv1.TupleKeyWithoutCondition {
-	return &authzextv1.TupleKeyWithoutCondition{
-		User:     t.GetUser(),
-		Relation: t.GetRelation(),
-		Object:   t.GetObject(),
-	}
-}
-
-func ToAuthzExtTupleKeysWithoutCondition(tuples []*openfgav1.TupleKeyWithoutCondition) []*authzextv1.TupleKeyWithoutCondition {
-	result := make([]*authzextv1.TupleKeyWithoutCondition, 0, len(tuples))
-	for _, t := range tuples {
-		result = append(result, ToAuthzExtTupleKeyWithoutCondition(t))
-	}
-	return result
-}
-
-func ToOpenFGATupleKey(t *authzextv1.TupleKey) *openfgav1.TupleKey {
-	tupleKey := &openfgav1.TupleKey{
-		User:     t.GetUser(),
-		Relation: t.GetRelation(),
-		Object:   t.GetObject(),
-	}
-
-	if t.GetCondition() != nil {
-		tupleKey.Condition = &openfgav1.RelationshipCondition{
-			Name:    t.GetCondition().GetName(),
-			Context: t.GetCondition().GetContext(),
-		}
-	}
-
-	return tupleKey
-}
-
-func ToOpenFGATupleKeyWithoutCondition(t *authzextv1.TupleKeyWithoutCondition) *openfgav1.TupleKeyWithoutCondition {
-	return &openfgav1.TupleKeyWithoutCondition{
-		User:     t.GetUser(),
-		Relation: t.GetRelation(),
-		Object:   t.GetObject(),
-	}
-}
-
-func ToOpenFGATuple(t *authzextv1.Tuple) *openfgav1.Tuple {
-	return &openfgav1.Tuple{
-		Key:       ToOpenFGATupleKey(t.GetKey()),
-		Timestamp: t.GetTimestamp(),
-	}
-}
-
-func ToOpenFGATuples(tuples []*authzextv1.Tuple) []*openfgav1.Tuple {
-	result := make([]*openfgav1.Tuple, 0, len(tuples))
-	for _, t := range tuples {
-		result = append(result, ToOpenFGATuple(t))
-	}
-	return result
 }
